requirements.txt
fastapi==0.63.0
uvicorn==0.13.4
mypy==0.812                                                                                                      
mypy-extensions==0.4.3
typing-extensions==3.7.4.3
websocket-client==0.57.0                                                                                          
websockets==8.1
tortoise-orm[aiosqlite]
pytest
email-validator